Unique Online Shopping and Delivery Application
I am looking for a developer to develop an application that can be run on Android, ioS and windows PC that will achieve business objectives of: shopping from any signed-up retailer, track progress of order execution, indicate progress of delivery and location of rider/driver on the way to customer. It should have functionality for display of products and prices, shopping cart and payment on delivery as well as in advance by mobile payment, credit, debit cards. 

Reach me on mwangangi.jimmi@gmail.com incase of any clarification required.